wadnr - Data hotfix - redundant Dals.
Fixing data error found in Prod(?)/QA, where we somehow had redundant Dal Marsters, which should not be allowed. Error dump below.

This fixes the data error, and adds a constraint to ensure it won't happen again in the data, but I would expect that whatever caused it will likely happen again, and I don't yet know what that is. So, stay tuned for more problems.